+++ xtp/src/libxtp/davidsonsolver.cc
@@ -92,9 +92,9 @@ void DavidsonSolver::printIterationData(
   double percent_converged = 100 * double(converged_roots) / double(neigen);
   XTP_LOG(Log::error, log_)
       << TimeStamp()
-      << boost::format(" %1$4d %2$12d \t %3$4.2e \t %4$5.2f%% converged") % i_iter_ %
-             proj.search_space() % rep.res_norm().head(neigen).maxCoeff() %
-             percent_converged
+      << boost::format(" %1$4d %2$12d \t %3$4.2e \t %4$5.2f%% converged") %
+             i_iter_ % proj.search_space() %
+             rep.res_norm().head(neigen).maxCoeff() % percent_converged
       << std::flush;
 }